♻️ Extract fields.go
